requirements.txt
pydantic
paho-mqtt
typing
PyYAML
influxdb-client
pandas
